+++ DAO_Programs/programs/dao/src/contexts/initialize.rs
@@ -6,6 +6,7 @@ use anchor_spl::{
     metadata::{Metadata, MetadataAccount,MasterEditionAccount}, 
     associated_token::AssociatedToken
 };
+use daoist_programs::modules::{StakeState, StakingProgram};
 use crate::{validate_nft, REQUIRED_COLLECTION_MINT};
 use crate::errors::DaoError;
 
@@ -79,17 +80,32 @@ impl<'info> Initialize<'info> {
         bumps: &InitializeBumps,
         proposal_fee: u64,
         min_quorum: u8,//0-100
+        //Minimum vote threshold
         min_threshold: u64,
+        //Maximum expiry time for proposals
         max_expiry: u64,
+        //Evaluation phase period for proposals
         evaluation_phase_period: u64,
         proposal_program: Pubkey,
         voting_program: Pubkey,
         staking_program: Pubkey,
+        //Specifies the mint of the NFT collection associated with the DAO
+        //Required for hybrid and NFT DAO
         collection_mint: Option<Pubkey>,
+        //Specifies the mint of the fungible token associated with the DAO.
+        //Required for hybrid and FT DAO
         mint: Option<Pubkey>,
+        //Minimum staked FT/NFT required to create a proposal
+        //Required for FT DAO & NFT DAO 
         min_staked_required_proposal: Option<u64>,
+        //Indicates whether sub-DAOs are allowed.
+        //set false for ARMADA
         allow_sub_dao: bool,
+        // Specifies the minimum stake required to create a sub-DAO.
+        //Required for FT DAO & NFT DAO
         min_staked_create_subdao: Option<u64>,
+        //Indicates whether the DAO is a hybrid type
+        //set true if hybrid
         is_hybrid: bool
     ) -> Result<()> {
 
@@ -127,8 +143,9 @@ impl<'info> Initialize<'info> {
     }
 }
 
+//Hybrid FLEETS - SUB FLEETS
 #[derive(Accounts)]
-#[instruction(name: String, seed: u64)]
+#[instruction(seed: u64)]
 pub struct InitializeSubdao<'info> {
     #[account(mut)]
     owner: Signer<'info>,
@@ -235,6 +252,93 @@ impl<'info> InitializeSubdao<'info> {
                 collection_mint,
                 mint,
                 min_staked_required_proposal,
+                allow_sub_dao: false,
+                min_staked_create_subdao: None,
+                is_hybrid
+            });
+                Ok(()) 
+   
+    }
+}
+
+// NFT/FT FLEETS - SUB FLEETS
+#[derive(Accounts)]
+#[instruction(seed: u64)]
+pub struct InitializeSubdaoToken<'info> {
+    #[account(mut)]
+    owner: Signer<'info>,
+    #[account(
+        seeds=[b"auth", config.key().as_ref()],
+        bump = config.auth_bump
+    )]
+    ///CHECK: This is safe. It's just used to sign things
+    auth: UncheckedAccount<'info>,
+    #[account(
+        seeds=[b"treasury", config_sub_dao.key().as_ref()],
+        bump
+    )]
+    treasury: SystemAccount<'info>,
+    #[account(
+        init,
+        payer = owner,
+        seeds=[b"config", seed.to_le_bytes().as_ref(), config.key().as_ref()],
+        bump,
+        space = DaoConfig::LEN
+    )]
+    config_sub_dao: Account<'info, DaoConfig>,
+    #[account(
+        seeds=[b"config", config.seed.to_le_bytes().as_ref()],
+        bump = config.config_bump
+    )]
+    config: Account<'info, DaoConfig>,
+    #[account(constraint = staking_program.key() == config.staking_program)]
+    staking_program: Program<'info, StakingProgram>,
+    #[account(
+        seeds=[b"stake", config_sub_dao.key().as_ref(), owner.key().as_ref()],
+        seeds::program = staking_program.key(),
+        bump = stake_state.state_bump,
+    )]
+    stake_state: Account<'info, StakeState>,
+    system_program: Program<'info, System>
+}
+
+impl<'info> InitializeSubdaoToken<'info> {
+    pub fn init(
+        &mut self, 
+        seed: u64,
+        bumps: &InitializeSubdaoTokenBumps,
+        proposal_fee: u64,
+        min_quorum: u8,//0-100
+        min_threshold: u64,
+        max_expiry: u64,
+        evaluation_phase_period: u64,
+        collection_mint: Option<Pubkey>,
+        mint: Option<Pubkey>,
+        min_staked_required_proposal : u64,
+        is_hybrid: bool
+    ) -> Result<()> {
+        
+            self.config.check_allow_sub_dao()?;
+            self.config.check_min_staked_create_subdao(self.stake_state.amount)?;
+            self.config_sub_dao.check_init_valid_quorum(min_quorum)?;       
+            self.config_sub_dao.set_inner(DaoConfig 
+                { 
+                seed, 
+                proposal_fee, 
+                min_quorum, 
+                min_threshold, 
+                max_expiry, 
+                evaluation_phase_period, 
+                proposal_count: 0, 
+                proposal_program: self.config.proposal_program, 
+                voting_program: self.config.voting_program, 
+                staking_program: self.config.staking_program, 
+                auth_bump: self.config.auth_bump, 
+                config_bump: bumps.config_sub_dao, 
+                treasury_bump: bumps.treasury,
+                collection_mint, 
+                mint,
+                min_staked_required_proposal: Some(min_staked_required_proposal),
                 allow_sub_dao: false,
                 min_staked_create_subdao: None,
                 is_hybrid
