+++ DAO_Programs/programs/dao/src/contexts/mod.rs
@@ -1,7 +1,5 @@
 pub mod initialize;
 pub mod core_handler;
-pub mod initialize_subdao;
 
-pub use initialize_subdao::*;
 pub use initialize::*;
 pub use core_handler::*;
