+++ DAO_Programs/programs/staking/src/contexts/cleanup_stake.rs
@@ -89,7 +89,7 @@ impl<'info> CleanupStake<'info> {
 
 
 }
-/* #[derive(Accounts)]
+#[derive(Accounts)]
 #[instruction(seed: u64)]
 pub struct CleanupStakeNft<'info> {
     #[account(mut)]
@@ -175,4 +175,3 @@ impl<'info> CleanupStakeNft<'info> {
 
 
 }
- */
